rename block effects resolver classes
